src/agent/requirements.txt
bcrypt
PyJWT
python-dotenv